+++ rollup-interface/src/lib.rs
@@ -3,7 +3,6 @@
 //! DA layers and be proven with different zkVMS, all while retaining compatibility
 //! with the same basic full node implementation.
 
-#![cfg_attr(not(feature = "std"), no_std)]
 #![deny(missing_docs)]
 
 extern crate alloc;
@@ -18,16 +17,3 @@ mod node;
 #[cfg(feature = "native")]
 pub use node::*;
 pub use {anyhow, digest};
-
-/// A facade for the `std` crate.
-pub mod maybestd {
-    #[cfg(not(target_has_atomic = "ptr"))]
-    pub use alloc::rc::Rc as RefCount;
-    // sync will be available only when the target supports atomic operations
-    #[cfg(target_has_atomic = "ptr")]
-    pub use alloc::sync;
-    #[cfg(target_has_atomic = "ptr")]
-    pub use alloc::sync::Arc as RefCount;
-
-    pub use borsh::maybestd::{borrow, boxed, collections, format, io, string, vec};
-}